requirements.txt
pandas
numpy
google-genai
python-dotenv
streamlit
watchdog
